WX-844-845 moved run details, reverted Call Caching text in CallTable
